Bob {RJSONIO}	R Documentation

Symbolic constants identifying the type of a JSON value.

Description

These constants are used by handler functions that are called when a JSON value is encountered by the JSON parser. These identify the type of the JSON value. The values will already have been converted, but the start and end array and object events won't have a type.

Format

A collection of integer values.

Source

JSON_parser.h code from http://www.json.org.

References

http://www.json.org.

[Package RJSONIO version 1.3-1.3 Index]